chat_app/api/requirements.txt
python-dotenv==0.8.2
pusher==2.0.1
SQLAlchemy==1.2.0
Flask-SQLAlchemy==2.1
Flask-JWT-Extended==3.13.0